Refactor according to feedback.
- Change index.jsx name to file name.
- Change api methods file name for each component.
- Fix variable declarations.
- Fix unnecessary new lines during imports.
-
