+++ src/Pages/Dashboard/Conversations/Conversations.jsx
@@ -1,12 +1,10 @@
 import { useState } from "react";
 import { useDispatch } from "react-redux";
-
 import { Conversation, ConversationList } from "@chatscope/chat-ui-kit-react";
 
 import { changeChatUser, setMessages } from "../../../Stores/Actions/chat";
 
-import { fetchUserMessageHistory } from "../Api/Methods";
-
+import { fetchUserMessageHistory } from "../Api/DashboardMethods";
 import { convertApiMessagesToChatMessages } from "../DashboardHelpers";
 
 const Conversations = ({ currentUser, users }) => {
