+++ src/Pages/Registration/Registration.jsx
@@ -1,4 +1,5 @@
 import { Link } from "react-router-dom";
+
 import RegistrationForm from "./Form/RegistrationForm";
 
 const Registration = () => {
